refactor: reorganize GameManager and introduce GameStates for better game state management
